backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
numpy==1.26.4
scipy==1.13.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
scikit-image==0.24.0
